requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
imagehash==4.3.1
numpy==1.26.2
piexif==1.1.3

